requirements.txt
fastapi
uvicorn
pydantic
openai==1.61.0
mem0ai